+++ src/asp/Camera/BundleAdjustCamera2.cc
@@ -62,7 +62,6 @@ void put_adjustments_in_params(std::string const& input_prefix,
     std::string adjust_file
       = asp::bundle_adjust_file_name(input_prefix, image_files[icam], camera_files[icam]);
   
-    vw_out() << "Reading input adjustment: " << adjust_file << std::endl;
     double * cam_ptr = param_storage.get_camera_ptr(icam);
     CameraAdjustment adjustment;
     adjustment.read_from_adjust_file(adjust_file);